use sweet alert

+++ src/app/Components/brand/brand.component.ts
@@ -4,6 +4,7 @@ import { Brand } from '../../Interfaces/Brand';
 import { NgClass, NgFor } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { RoleDirective } from '../../Directive/role.directive';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-brand',
@@ -25,10 +26,36 @@ export class BrandComponent implements OnInit {
   }
 
   deleteBrand(id:number){
-    this.brandService.deleteBrand(id).subscribe(()=>{
-      this.brands=this.brands.filter(p=>p.id !== id);
+     Swal.fire({
+              title: "Are you sure?",
+              text: "You won't be able to revert this!",
+              icon: "warning",
+              showCancelButton: true,
+              confirmButtonColor: "#3085d6",
+              cancelButtonColor: "#d33",
+              confirmButtonText: "Yes, delete it!"
+            }).then((result) => {
+              if (result.isConfirmed) {
+                this.brandService.deleteBrand(id).subscribe(() => {
+             
+                  this.brands = this.brands.filter(p => p.id !== id); 
+                  Swal.fire({
+                    title: "Deleted!",
+                    text: "Deleted successfully.",
+                    icon: "success"
+                  });
+                }, error => {
+                  Swal.fire({
+                    title: "Error!",
+                    text: "Something went wrong.",
+                    icon: "error"
+                  });
+                });
+              }
+            });
+
   
-    });
+
   }
   sortByCreationDate(sortDirection:string){
     this.brandService.sortByCreationDate(sortDirection).subscribe(data=>this.brands=data);
