+++ src/app/Components/category/category.component.ts
@@ -4,6 +4,7 @@ import { CategoryService } from '../../Services/category.service';
 import { NgFor } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { RoleDirective } from '../../Directive/role.directive';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-category',
@@ -22,10 +23,36 @@ ngOnInit(): void {
 }
 
 deleteCategory(id:number){
-  this.categoryService.deleteCategory(id).subscribe(()=>{
-    this.category=this.category.filter(p=>p.id !== id);
+   Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.categoryService.deleteCategory(id).subscribe(() => {
+         
+              this.category = this.category.filter(p => p.id !== id); 
+              Swal.fire({
+                title: "Deleted!",
+                text: "Deleted successfully.",
+                icon: "success"
+              });
+            }, error => {
+              Swal.fire({
+                title: "Error!",
+                text: "Something went wrong.",
+                icon: "error"
+              });
+            });
+          }
+        });
+  
 
-  });
+ 
 }
 sortByCreationDate(sortDirection:string){
   this.categoryService.sortByCreationDate(sortDirection).subscribe(data=>this.category=data);
