+++ src/app/Components/product-table/product-table.component.ts
@@ -4,6 +4,7 @@ import { Router, RouterLink } from '@angular/router';
 import { Product } from '../../Interfaces/Product';
 import { CurrencyPipe, NgFor } from '@angular/common';
 import { RoleDirective } from '../../Directive/role.directive';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product-table',
@@ -37,9 +38,36 @@ this.productService.getProducts().subscribe(data=>this.products=data);
     
     }
     deleteProduct(id:number){
-      this.productService.deleteProduct(id).subscribe(()=>{
-        this.products=this.products.filter(p=>p.id !== id);
-    });
+      Swal.fire({
+                    title: "Are you sure?",
+                    text: "You won't be able to revert this!",
+                    icon: "warning",
+                    showCancelButton: true,
+                    confirmButtonColor: "#3085d6",
+                    cancelButtonColor: "#d33",
+                    confirmButtonText: "Yes, delete it!"
+                  }).then((result) => {
+                    if (result.isConfirmed) {
+                      this.productService.deleteProduct(id).subscribe(() => {
+                   
+                        this.products = this.products.filter(p => p.id !== id); 
+                        Swal.fire({
+                          title: "Deleted!",
+                          text: "Deleted successfully.",
+                          icon: "success"
+                        });
+                      }, error => {
+                        Swal.fire({
+                          title: "Error!",
+                          text: "Something went wrong.",
+                          icon: "error"
+                        });
+                      });
+                    }
+                  });
+      
+        
+
   }
  
 }
