+++ src/app/Components/roles/roles.component.ts
@@ -4,6 +4,7 @@ import { Role } from '../../Interfaces/Role';
 import { NgFor } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { RoleDirective } from '../../Directive/role.directive';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-roles',
@@ -21,12 +22,35 @@ export class RolesComponent implements OnInit {
   ngOnInit(): void {
     this.roleService.getRoles().subscribe(data=>this.roles=data);
   }
-  deleteRole(id:number){
-    this.roleService.deleteRole(id).subscribe(()=>{
-      this.roles=this.roles.filter(p=>p.id !== id);
-  
-    });
-  }
+  deleteRole(id: number) {
+  Swal.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Yes, delete it!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.roleService.deleteRole(id).subscribe(() => {
+   
+        this.roles = this.roles.filter(p => p.id !== id); 
+        Swal.fire({
+          title: "Deleted!",
+          text: "Deleted successfully.",
+          icon: "success"
+        });
+      }, error => {
+        Swal.fire({
+          title: "Error!",
+          text: "Something went wrong.",
+          icon: "error"
+        });
+      });
+    }
+  });
+}
 
     sortByCreationDate(sortDirection:string){
       this.roleService.sortByCreationDate(sortDirection).subscribe(data=>this.roles=data);
