+++ src/app/Components/subcategory-table/subcategory-table.component.ts
@@ -4,6 +4,7 @@ import { NgFor } from '@angular/common';
 import { SubCategory } from '../../Interfaces/SubCategory';
 import { SubcategoryService } from '../../Services/subcategory.service';
 import { RouterLink } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-subcategory-table',
@@ -21,10 +22,35 @@ ngOnInit(): void {
 }
 
 deleteSubCategory(id:number){
-  this.subCategoryService.deleteSubCategory(id).subscribe(()=>{
-    this.category=this.category.filter(p=>p.id !== id);
+   Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.subCategoryService.deleteSubCategory(id).subscribe(() => {
+     
+          this.category = this.category.filter(p => p.id !== id); 
+          Swal.fire({
+            title: "Deleted!",
+            text: "Deleted successfully.",
+            icon: "success"
+          });
+        }, error => {
+          Swal.fire({
+            title: "Error!",
+            text: "Something went wrong.",
+            icon: "error"
+          });
+        });
+      }
+    });
 
-  });
+
 }
 sortByCreationDate(sortDirection:string){
   this.subCategoryService.sortByCreationDate(sortDirection).subscribe(data=>this.category=data);
