+++ src/app/Components/user/user.component.ts
@@ -6,6 +6,7 @@ import { RouterLink } from '@angular/router';
 import { RoleService } from '../../Services/role.service';
 import { Role } from '../../Interfaces/Role';
 import { RoleDirective } from '../../Directive/role.directive';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -35,10 +36,34 @@ export class UserComponent implements OnInit {
   
   
   deleteUser(id:number){
-    this.userService.deleteUser(id).subscribe(()=>{
-      this.users=this.users.filter(p=>p.id !== id);
-  
-    });
+     Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.userService.deleteUser(id).subscribe(() => {
+       
+            this.users = this.users.filter(p => p.id !== id); 
+            Swal.fire({
+              title: "Deleted!",
+              text: "Deleted successfully.",
+              icon: "success"
+            });
+          }, error => {
+            Swal.fire({
+              title: "Error!",
+              text: "Something went wrong.",
+              icon: "error"
+            });
+          });
+        }
+      });
+
   }
   sortByCreationDate(sortDirection:string){
     this.userService.sortByCreationDate(sortDirection).subscribe(data=>this.users=data);
